feat: bottom navigation bar

+++ lib/constants/app_dimensions.dart
@@ -1,3 +1,3 @@
 class AppDimensions {
-  static const int defaultPadding = 25;
+  static const double defaultPadding = 25;
 }
